Refactor timetable to use epochs and availability data
A substantial refactoring was done on the timetable feature. Earlier the system was operating using AgendaItems to function. However, its functionality has been modified to use epoch time and availability information represented as AvailabilityItems. The change was necessitated to better cater to the requirements and improve the performance. The AvailabilityItem now hosts the start and end time as long values representing seconds from epoch and a status.

The timetable also now fetches the agenda for the current week by making use of the readAvailability method. Similarly, additional changes relate to updating various existing methods to align with the new AvailabilityItem use. The DaySelector was also accordingly adjusted to correctly handle this transformation.

Filenames and some identifiers were adjusted to ensure consistency and improve readability. It is now easier to read and understand what each file or identifier signifies, making future debugging and development smoother. The changes were carefully implemented to not hamper the existing functionalities while improving the timetable feature's overall performance and effectiveness.