Modify API URLs to be fetched dynamically
AvailabilityUrl, requestHoursUrl, and infoUrl in the API were previously static URLs, which caused problems when the personId changed, as the URLs did not update to reflect this change. Now, these URLs are made dynamic using the 'get' keyword, which ensures that they are updated anytime the personId changes, improving application reliability. This change was prompted by issues with the old system not accurately reflecting changes in personId.