Introduce error messages in login screens
The change involves adding error messages on LoginScreen.kt and App.kt. In the LoginScreen, we now display error messages for incorrect usernames/passwords and issues during Magister sign in, improving user feedback for failed login attempts. The @OptIn annotation was also removed from the App function as it was deemed unnecessary. Additionally, two new imports for Settings and Icons were added in App.kt, which may be utilized in future commits.